Move time related processes into a separate module
