+++ src/main/scala/scalaz/stream/Process.scala
@@ -1,10 +1,8 @@
 package scalaz.stream
 
 import Cause._
-import java.util.concurrent.ScheduledExecutorService
 import scala.annotation.tailrec
 import scala.collection.SortedMap
-import scala.concurrent.duration._
 import scalaz.{Catchable, Functor, Monad, Monoid, Nondeterminism, \/, -\/, ~>}
 import scalaz.\/._
 import scalaz.concurrent.{Actor, Strategy, Task}
@@ -777,49 +775,6 @@ object Process extends ProcessInstances {
     liftW(Process.awaitBoth[I, I2])
 
   /**
-   * Discrete process that every `d` emits elapsed duration
-   * since the start time of stream consumption.
-   *
-   * For example: `awakeEvery(5 seconds)` will
-   * return (approximately) `5s, 10s, 20s`, and will lie dormant
-   * between emitted values.
-   *
-   * By default, this uses a shared `ScheduledExecutorService`
-   * for the timed events, and runs the consumer using the `pool` `Strategy`,
-   * to allow for the process to decide whether result shall be run on
-   * different thread pool, or with `Strategy.Sequential` on the
-   * same thread pool as the scheduler.
-   *
-   * @param d           Duration between emits of the resulting process
-   * @param S           Strategy to run the process
-   * @param scheduler   Scheduler used to schedule tasks
-   */
-  def awakeEvery(d: Duration)(
-    implicit S: Strategy,
-    scheduler: ScheduledExecutorService): Process[Task, Duration] = {
-    def metronomeAndSignal:(()=>Unit,async.mutable.Signal[Duration]) = {
-      val t0 = Duration(System.nanoTime, NANOSECONDS)
-      val signal = async.toSignal[Duration](Process.halt)(S)
-
-      val metronome = scheduler.scheduleAtFixedRate(
-        new Runnable { def run = {
-          val d = Duration(System.nanoTime, NANOSECONDS) - t0
-          signal.set(d).run
-        }},
-        d.toNanos,
-        d.toNanos,
-        NANOSECONDS
-      )
-      (()=>metronome.cancel(false), signal)
-    }
-
-    await(Task.delay(metronomeAndSignal))({
-      case (cm, signal) =>  signal.discrete onComplete eval_(signal.close.map(_=>cm()))
-    })
-  }
-
-
-  /**
    * The infinite `Process`, always emits `a`.
    * If for performance reasons it is good to emit `a` in chunks,
    * specify size of chunk by `chunkSize` parameter
@@ -831,16 +786,6 @@ object Process extends ProcessInstances {
     go
   }
 
-  /**
-   * A continuous stream of the elapsed time, computed using `System.nanoTime`.
-   * Note that the actual granularity of these elapsed times depends on the OS, for instance
-   * the OS may only update the current time every ten milliseconds or so.
-   */
-  def duration: Process[Task, FiniteDuration] =
-    eval(Task.delay(System.nanoTime)).flatMap { t0 =>
-      repeatEval(Task.delay(FiniteDuration(System.nanoTime - t0, NANOSECONDS)))
-    }
-
   /** A `Writer` which emits one value to the output. */
   def emitO[O](o: O): Process0[Nothing \/ O] =
     Process.emit(right(o))
@@ -848,22 +793,6 @@ object Process extends ProcessInstances {
   /** A `Writer` which writes the given value. */
   def emitW[W](s: W): Process0[W \/ Nothing] =
     Process.emit(left(s))
-
-  /**
-   * A 'continuous' stream which is true after `d, 2d, 3d...` elapsed duration,
-   * and false otherwise.
-   * If you'd like a 'discrete' stream that will actually block until `d` has elapsed,
-   * use `awakeEvery` instead.
-   */
-  def every(d: Duration): Process[Task, Boolean] = {
-    def go(lastSpikeNanos: Long): Process[Task, Boolean] =
-      suspend {
-        val now = System.nanoTime
-        if ((now - lastSpikeNanos) > d.toNanos) emit(true) ++ go(now)
-        else emit(false) ++ go(lastSpikeNanos)
-      }
-    go(0)
-  }
 
   /** A `Process` which emits `n` repetitions of `a`. */
   def fill[A](n: Int)(a: A, chunkSize: Int = 1): Process0[A] = {
@@ -934,19 +863,6 @@ object Process extends ProcessInstances {
           None
     }
   }
-
-  /**
-   * A single-element `Process` that waits for the duration `d`
-   * before emitting its value. This uses a shared
-   * `ScheduledThreadPoolExecutor` to signal duration and
-   * avoid blocking on thread. After the signal,
-   * the execution continues with `S` strategy
-   */
-  def sleep(d: FiniteDuration)(
-    implicit S: Strategy
-    , schedulerPool: ScheduledExecutorService
-    ): Process[Task, Nothing] =
-    awakeEvery(d).once.drain
 
   /**
    * Delay running `p` until `awaken` becomes true for the first time.
